graphics: use better types for position and color buffers

+++ src/graphics/buffer.h
@@ -48,8 +48,8 @@ struct Buffer {
 
 	unsigned int size;
 	GLuint buffers[4]; // first is for positions, second for colors, third (optional) for texcoords, forth (optional) for point sizes
-	GLfloat* positions;
-	GLfloat* colors;
+	GLshort* positions;
+	GLubyte* colors;
 	GLfloat* tex_coords; // only if has_texture
 	GLfloat* point_sizes; // only if POINT_BUFFER
 	unsigned int current_position;
@@ -71,8 +71,8 @@ Buffer *buffer_new(bool user_buffer, unsigned int size);
 void buffer_free(Buffer *b);
 void buffer_allocate(Buffer *b);
 
-void buffer_push_vertex(Buffer *b, GLfloat, GLfloat);
-void buffer_push_color(Buffer *b, GLfloat, GLfloat, GLfloat, GLfloat);
+void buffer_push_vertex(Buffer *b, GLshort, GLshort);
+void buffer_push_color(Buffer *b, GLubyte, GLubyte, GLubyte, GLubyte);
 void buffer_push_tex_coord(Buffer *b, GLfloat, GLfloat);
 void buffer_push_point_size(Buffer *b, GLfloat);
